+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/node/metrics/MetricsService.java
@@ -64,6 +64,7 @@
 package com.radixdlt.api.node.metrics;
 
 import com.radixdlt.api.node.account.AccountInfoService;
+import com.radixdlt.api.service.NetworkingService;
 import com.radixdlt.api.service.SystemConfigService;
 import com.radixdlt.api.Endpoints;
 import com.radixdlt.api.service.network.NetworkInfoService;
@@ -131,6 +132,7 @@ public class MetricsService {
 	private final SystemCounters systemCounters;
 	private final InfoSupplier infoSupplier;
 	private final SystemConfigService systemConfigService;
+	private final NetworkingService networkingService;
 	private final AccountInfoService accountInfoService;
 	private final ValidatorInfoService validatorInfoService;
 	private final NetworkInfoService networkInfoService;
@@ -145,6 +147,7 @@ public class MetricsService {
 		SystemCounters systemCounters,
 		InfoSupplier infoSupplier,
 		SystemConfigService systemConfigService,
+		NetworkingService networkingService,
 		AccountInfoService accountInfoService,
 		ValidatorInfoService validatorInfoService,
 		NetworkInfoService networkInfoService,
@@ -156,6 +159,7 @@ public class MetricsService {
 		this.systemCounters = systemCounters;
 		this.infoSupplier = infoSupplier;
 		this.systemConfigService = systemConfigService;
+		this.networkingService = networkingService;
 		this.accountInfoService = accountInfoService;
 		this.validatorInfoService = validatorInfoService;
 		this.networkInfoService = networkInfoService;
@@ -179,7 +183,7 @@ public class MetricsService {
 		appendCounter(builder, "info_configuration_pacemakermaxexponent", pacemakerMaxExponent(snapshot));
 		appendCounter(builder, "info_epochmanager_currentview_view", currentView(snapshot));
 		appendCounter(builder, "info_epochmanager_currentview_epoch", currentEpoch(snapshot));
-		appendCounter(builder, "total_peers", systemConfigService.getNetworkingPeersCount());
+		appendCounter(builder, "total_peers", networkingService.getPeersCount());
 
 		var totalValidators = inMemorySystemInfo.getEpochProof().getNextValidatorSet()
 			.map(BFTValidatorSet::getValidators)
